feat(webpack): Change webpack loading stratgies and implement production mode
Check for the prescence of the `-p` flag on the command line to determine whether webpack is running
in production mode or not. Only enable hot module loading and named modules when not in production
mode. Also removes vendors as an explicit entry point and uses CommonsChunkPlugin to force all
modules in `node_modules` into a separate chunk.

`entries` now contains a single entry point, (./src/index.tsx) - vendors are configured
automatically using a new CommonsChunkPlugin configuration