feat(ResourceTable): Implement `customTitle` function prop
Allow `ResourceTable` to accept a `customTitle` prop, which functions in the same way as the `title`
prop on the Ant Design Table component, except that the currently fetched resource is also supplied
as a second parameter. Allows table titles to be constructed using meta-properties on the top-level
resource, for example.